# webapp/public/swagger/swagger-v2.json
{
  "swagger": "2.0",
  "info": {
    "description": "DT Paths API.",
    "version": "0.0.1",
    "title": "DT Paths",
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    }
  },
  "basePath": "/api",
  "schemes": [
    "http"
  ],
  "paths": {
    "/dt": {
      "get": {
        "summary": "Show DTs",
        "description": "Show DTs.",
        "operationId": "dt",
        "produces": [
          "application/json"
        ],
        "parameters": [ ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "type": "array",
              "items": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "available": {
                    "type": "boolean"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Error"
          }
        }
      }
    },
    "/sim": {
      "get": {
        "summary": "Query DT",
        "description": "Query DT.",
        "operationId": "sim",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "description": "The search query item.",
            "required": true,
            "default": "Jaguar#NN",
            "type": "string"
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Limit the query results.",
            "required": false,
            "default": 200,
            "type": "integer"
          },
          {
            "name": "offset",
            "in": "query",
            "description": "Skip query results.",
            "required": false,
            "default": 1,
            "type": "integer"
          },
          {
            "name": "dt",
            "in": "query",
            "description": "The DT to use.",
            "required": true,
            "default": "default",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "type": "array",
              "items": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "weight": {
                    "type": "number",
                    "format": "float"
                  },
                  "rank": {
                    "type": "integer",
                    "format": "int32"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Error"
          }
        }
      }
    },
    "/path": {
      "get": {
        "summary": "Get path within DT.",
        "description": "Get path within DT.",
        "operationId": "path",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "start",
            "in": "query",
            "description": "The start query item.",
            "required": true,
            "type": "string",
            "default": "Jaguar#NN"
          },
          {
            "name": "dest",
            "in": "query",
            "description": "The destination query item.",
            "required": true,
            "type": "string",
            "default": "tiger#NN"
          },
          {
            "name": "topk",
            "in": "query",
            "description": "The number of neighbors to generate.",
            "required": false,
            "default": 200,
            "type": "integer"
          },
          {
            "name": "dt",
            "in": "query",
            "description": "The DT to use.",
            "required": true,
            "default": "default",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "type": "object"
            }
          },
          "400": {
            "description": "Error"
          }
        }
      }
    }
  }
}